feat: support downloading other shows.
I removed the season and episode numbering in the filename because I
realized I was doing it wrong™. I'm not sure of the correct/exact
scheme PBS is using.

Filenames are a little yucky still at the moment. I remove the spare /
in the titles and replace them with dashes. They are still apostrophes
hanging around but those shouldn't cause any issues.

Specifying shows can be done with -S and by default seasame-street is
used so as to not break existing setups.

Bitrates have added a wrinkle. The bitrates vary show by show so in
this commit I add support for 4500k bitrate, that may be too high but
I'm not sure if 2000k is too low. 4500k seems to correspond to 1080p.